searchbar funcitonal for teams
