+++ src/components/SearchBar/SearchBar.jsx
@@ -2,8 +2,7 @@ import React, { useState } from 'react'
 import style from './SearchBar.module.css'
 import lupa from '../../assets/lupa.png'
 import { useDispatch }from 'react-redux'
-//crear action
-
+import { searchTeam } from '../../redux/actions/searchTeamLeague.js'
 
 const SearchBar = () => {
 
@@ -17,12 +16,13 @@ const SearchBar = () => {
   
   const handleSubmit = (e) => {
     e.preventDefault()
-    dispatch()
+    dispatch(searchTeam(search))
+    setSearch("")
   }
 
   return (
     <div className={style.searchBar}>
-    <input value={search}type="text" onChange={handleChange} className={style.searchInput} placeholder="Buscar equipo/liga..." />
+    <input value={search}type="text" onChange={handleChange} className={style.searchInput} placeholder="Buscar equipo" />
     <button onClick={handleSubmit} className={style.searchButton}>
       <img className={style.img} src={lupa} alt="Lupa" />
     </button>
