+++ src/index.js
@@ -5,9 +5,14 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 
+import axios from "axios";
+
 import { Provider } from 'react-redux';
 
 import store from './redux/store/store.js';
+
+
+axios.defaults.baseURL = 'http://localhost:3001'
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
